move get element props helper
