--- lib/shared/helpers/get-element-props.js
@@ -1,18 +1,0 @@
-import displays from 'statics/displays';
-import forEach from 'lodash/forEach';
-
-export default function getElementProps (element, display) {
-  let elementProps = element.props || {};
-
-  if (display !== 'desktop' && element.displayProps) {
-    const changes = {};
-    forEach(displays, (value, displayIt) => {
-      if (displayIt !== 'desktop' && value >= displays[display] && element.displayProps[displayIt]) {
-        Object.assign(changes, element.displayProps[displayIt]);
-      }
-    });
-    elementProps = Object.assign({}, elementProps, changes);
-  }
-
-  return elementProps;
-}
