+++ lib/shared/helpers/traverser/element.js
@@ -1,6 +1,6 @@
 import Symbol from 'elements/symbol';
 import elements from 'elements';
-import getElementProps from 'helpers/get-element-props';
+import getElementProps from 'helpers/element/get-props';
 
 import calculateLinks from './links';
